fix: swapper transition states
