+++ src/context/AppProvider/AppContext.tsx
@@ -1,11 +1,6 @@
 import { usePrevious, useToast } from '@chakra-ui/react'
-import type { AccountId, ChainId } from '@shapeshiftoss/caip'
-import { fromAccountId } from '@shapeshiftoss/caip'
 import type { LedgerOpenAppEventArgs } from '@shapeshiftoss/chain-adapters'
 import { emitter } from '@shapeshiftoss/chain-adapters'
-import { isLedger } from '@shapeshiftoss/hdwallet-ledger'
-import { MetaMaskShapeShiftMultiChainHDWallet } from '@shapeshiftoss/hdwallet-shapeshift-multichain'
-import type { AccountMetadataById } from '@shapeshiftoss/types'
 import { useQueries } from '@tanstack/react-query'
 import { DEFAULT_HISTORY_TIMEFRAME } from 'constants/Config'
 import { LanguageTypeEnum } from 'constants/LanguageTypeEnum'
@@ -19,8 +14,6 @@ import { useModal } from 'hooks/useModal/useModal'
 import { useRouteAssetId } from 'hooks/useRouteAssetId/useRouteAssetId'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { walletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
-import { deriveAccountIdsAndMetadata } from 'lib/account/account'
-import { isUtxoChainId } from 'lib/utils/utxo'
 import { snapshotApi } from 'state/apis/snapshot/snapshot'
 import { useGetAssetsQuery } from 'state/slices/assetsSlice/assetsSlice'
 import {
@@ -28,12 +21,11 @@ import {
   marketData,
   useFindAllQuery,
 } from 'state/slices/marketDataSlice/marketDataSlice'
-import { portfolio, portfolioApi } from 'state/slices/portfolioSlice/portfolioSlice'
+import { portfolio } from 'state/slices/portfolioSlice/portfolioSlice'
 import { preferences } from 'state/slices/preferencesSlice/preferencesSlice'
 import {
   selectAccountIdsByChainId,
   selectAssetIds,
-  selectEnabledWalletAccountIds,
   selectPortfolioAssetIds,
   selectPortfolioLoadingStatus,
   selectSelectedCurrency,
@@ -41,8 +33,9 @@ import {
   selectWalletId,
 } from 'state/slices/selectors'
 import { tradeInput } from 'state/slices/tradeInputSlice/tradeInputSlice'
-import { txHistoryApi } from 'state/slices/txHistorySlice/txHistorySlice'
 import { useAppDispatch, useAppSelector } from 'state/store'
+
+import { useAccountsFetchQuery } from './hooks/useAccountsFetchQuery'
 
 /**
  * note - be super careful playing with this component, as it's responsible for asset,
@@ -61,14 +54,12 @@ export const AppProvider = ({ children }: { children: React.ReactNode }) => {
   const { supportedChains } = usePlugins()
   const { wallet, isConnected } = useWallet().state
   const assetIds = useAppSelector(selectAssetIds)
-  const requestedAccountIds = useAppSelector(selectEnabledWalletAccountIds)
   const portfolioLoadingStatus = useAppSelector(selectPortfolioLoadingStatus)
   const portfolioAssetIds = useAppSelector(selectPortfolioAssetIds)
   const walletId = useAppSelector(selectWalletId)
   const prevWalletId = usePrevious(walletId)
   const routeAssetId = useRouteAssetId()
   const { isSnapInstalled } = useIsSnapInstalled()
-  const previousIsSnapInstalled = usePrevious(isSnapInstalled)
   const { close: closeModal, open: openModal } = useModal('ledgerOpenApp')
 
   useEffect(() => {
@@ -110,6 +101,9 @@ export const AppProvider = ({ children }: { children: React.ReactNode }) => {
   // and covers most assets users will have
   useFindAllQuery()
 
+  // Master hook for accounts fetch as a react-query
+  useAccountsFetchQuery()
+
   const selectedLocale = useAppSelector(selectSelectedLocale)
   useEffect(() => {
     if (selectedLocale in LanguageTypeEnum ?? {}) {
@@ -118,6 +112,7 @@ export const AppProvider = ({ children }: { children: React.ReactNode }) => {
   }, [selectedLocale])
 
   const accountIdsByChainId = useAppSelector(selectAccountIdsByChainId)
+
   useEffect(() => {
     if (!wallet) return
     const walletSupportedChainIds = supportedChains.filter(chainId => {
@@ -130,185 +125,6 @@ export const AppProvider = ({ children }: { children: React.ReactNode }) => {
     })
     dispatch(portfolio.actions.setWalletSupportedChainIds(walletSupportedChainIds))
   }, [accountIdsByChainId, dispatch, isSnapInstalled, wallet, supportedChains])
-
-  // Initial account and portfolio fetch for non-ledger wallets
-  useEffect(() => {
-    const hasManagedAccounts = (() => {
-      // MM without snap doesn't allow account management - if the user just installed the snap, we know they don't have managed accounts
-      if (!previousIsSnapInstalled && isSnapInstalled) return false
-      // We know snap wasn't just installed in this render - so if there are any requestedAccountIds, we assume the user has managed accounts
-      return requestedAccountIds.length > 0
-    })()
-
-    let chainIds = new Set(
-      supportedChains.filter(chainId => {
-        return walletSupportsChain({
-          chainId,
-          wallet,
-          isSnapInstalled,
-          checkConnectedAccountIds: false, // don't check connected account ids, we're detecting runtime support for chains
-        })
-      }),
-    )
-    if (!chainIds.size) return
-    ;(async () => {
-      dispatch(portfolio.actions.setIsAccountsMetadataLoading(true))
-
-      // Fetch portfolio for all managed accounts if they exist instead of going through the initial account detection flow.
-      // This ensures that we have fresh portfolio data, but accounts added through account management are not accidentally blown away.
-      if (hasManagedAccounts) {
-        requestedAccountIds.forEach(accountId => {
-          dispatch(portfolioApi.endpoints.getAccount.initiate({ accountId, upsertOnFetch: true }))
-        })
-
-        return
-      }
-
-      if (!wallet || isLedger(wallet)) return
-
-      const walletId = await wallet.getDeviceID()
-
-      const accountMetadataByAccountId: AccountMetadataById = {}
-      const isMultiAccountWallet = wallet.supportsBip44Accounts()
-      const isMetaMaskMultichainWallet = wallet instanceof MetaMaskShapeShiftMultiChainHDWallet
-      for (let accountNumber = 0; chainIds.size > 0; accountNumber++) {
-        if (
-          accountNumber > 0 &&
-          // only some wallets support multi account
-          (!isMultiAccountWallet ||
-            // MM without snaps does not support non-EVM chains, hence no multi-account
-            // since EVM chains in MM use MetaMask's native JSON-RPC functionality which doesn't support multi-account
-            (isMetaMaskMultichainWallet && !isSnapInstalled))
-        )
-          break
-
-        const input = {
-          accountNumber,
-          chainIds: Array.from(chainIds),
-          wallet,
-          isSnapInstalled: Boolean(isSnapInstalled),
-        }
-        const accountIdsAndMetadata = await deriveAccountIdsAndMetadata(input)
-        const accountIds = Object.keys(accountIdsAndMetadata)
-
-        Object.assign(accountMetadataByAccountId, accountIdsAndMetadata)
-
-        const { getAccount } = portfolioApi.endpoints
-
-        const accountNumberAccountIdsByChainId = (
-          _accountIds: AccountId[],
-        ): Record<ChainId, AccountId[]> => {
-          return _accountIds.reduce(
-            (acc, _accountId) => {
-              const { chainId } = fromAccountId(_accountId)
-
-              if (!acc[chainId]) {
-                acc[chainId] = []
-              }
-              acc[chainId].push(_accountId)
-
-              return acc
-            },
-            {} as Record<ChainId, AccountId[]>,
-          )
-        }
-
-        let chainIdsWithActivity: Set<ChainId> = new Set()
-        // This allows every run of AccountIds per chain/accountNumber to run in parallel vs. all sequentally, so
-        // we can run each item (usually one AccountId, except UTXOs which may contain many because of many scriptTypes) 's side effects immediately
-        const accountNumberAccountIdsPromises = Object.values(
-          accountNumberAccountIdsByChainId(accountIds),
-        ).map(async accountIds => {
-          const results = await Promise.allSettled(
-            accountIds.map(async id => {
-              const result = await dispatch(getAccount.initiate({ accountId: id }))
-              return result
-            }),
-          )
-
-          results.forEach((res, idx) => {
-            if (res.status === 'rejected') return
-
-            const { data: account } = res.value
-            if (!account) return
-
-            const accountId = accountIds[idx]
-            const { chainId } = fromAccountId(accountId)
-
-            const { hasActivity } = account.accounts.byId[accountId]
-
-            const accountNumberHasChainActivity = !isUtxoChainId(chainId)
-              ? hasActivity
-              : // For UTXO AccountIds, we need to check if *any* of the scriptTypes have activity, not only the current one
-                // else, we might end up with partial account data, with only the first 1 or 2 out of 3 scriptTypes
-                // being upserted for BTC and LTC
-                results.some((res, _idx) => {
-                  if (res.status === 'rejected') return false
-                  const { data: account } = res.value
-                  if (!account) return false
-                  const accountId = accountIds[_idx]
-                  const { chainId: _chainId } = fromAccountId(accountId)
-                  if (chainId !== _chainId) return false
-                  return account.accounts.byId[accountId].hasActivity
-                })
-
-            // don't add accounts with no activity past account 0
-            if (accountNumber > 0 && !accountNumberHasChainActivity) {
-              chainIdsWithActivity.delete(chainId)
-              delete accountMetadataByAccountId[accountId]
-            } else {
-              // handle utxo chains with multiple account types per account
-              chainIdsWithActivity.add(chainId)
-
-              dispatch(portfolio.actions.upsertPortfolio(account))
-              const chainIdAccountMetadata = Object.entries(accountMetadataByAccountId).reduce(
-                (acc, [accountId, metadata]) => {
-                  const { chainId: _chainId } = fromAccountId(accountId)
-                  if (chainId === _chainId) {
-                    acc[accountId] = metadata
-                  }
-                  return acc
-                },
-                {} as AccountMetadataById,
-              )
-              dispatch(
-                portfolio.actions.upsertAccountMetadata({
-                  accountMetadataByAccountId: chainIdAccountMetadata,
-                  walletId,
-                }),
-              )
-              for (const accountId of Object.keys(accountMetadataByAccountId)) {
-                dispatch(portfolio.actions.enableAccountId(accountId))
-              }
-            }
-          })
-
-          return results
-        })
-
-        await Promise.allSettled(accountNumberAccountIdsPromises)
-        chainIds = chainIdsWithActivity
-      }
-    })().then(async () => {
-      dispatch(portfolio.actions.setIsAccountsMetadataLoading(false))
-      // Only fetch and upsert Tx history once all are loaded, otherwise big main thread rug
-      const { getAllTxHistory } = txHistoryApi.endpoints
-
-      await Promise.all(
-        requestedAccountIds.map(requestedAccountId =>
-          dispatch(getAllTxHistory.initiate(requestedAccountId)),
-        ),
-      )
-    })
-  }, [
-    dispatch,
-    wallet,
-    supportedChains,
-    isSnapInstalled,
-    requestedAccountIds.length,
-    previousIsSnapInstalled,
-    requestedAccountIds,
-  ])
 
   useEffect(() => {
     if (portfolioLoadingStatus === 'loading') return
